Change the file

+++ src/Week1/Help.java
@@ -4,7 +4,7 @@ public class Help {
             System.out.println("    1. if");
             System.out.println("    2. switch");
             System.out.print("Choose one: ");
-            int input = (int)System.in.read();
+            int input = System.in.read();
         switch(input) {
             case 1:
                 System.out.println("The if:");
